restructured files and fixed css
