+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@availity/favorites": "^1.1.9",
     "@sentry/browser": "^5.5.0",
     "axios": "^0.19.0",
     "enzyme": "^3.10.0",
