+++ src/views/Home/Home.jsx
@@ -1,14 +1,11 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import axios from 'axios';
-import track from '../../utils/analytics';
-import store from '../../utils/jwt-store';
-import feature from '../../utils/flaggie';
-import styles from '../../App.module.scss';
+import PropTypes from 'prop-types';
+import styles from './Home.module.scss';
 import Search from '../../components/Search';
 import PopularDestinations from '../../components/PopularDestinations';
 import Attractions from '../../components/Attractions';
 import Hero from '../../components/Hero';
-import Modal from '../../components/Modal';
 
 axios.interceptors.request.use((config) => {
   const newConfig = {
@@ -22,65 +19,14 @@ axios.interceptors.request.use((config) => {
   return newConfig;
 });
 
-function Home() {
+function Home({ showModal }) {
   const [state, setState] = useState({
     clientId: process.env.REACT_APP_OAUTH_GOOGLE_ID,
     attractions: [],
     destination: '',
     isLoading: false,
-    modal: {
-      show: false,
-      attraction: {},
-    },
     noResults: false,
   });
-
-  const [features, setFeatures] = useState({
-    loading: true,
-    flags: {},
-  });
-
-  const Feature = feature(features.flags, features.loading);
-
-  // For loading feature flags
-  useEffect(() => {
-    const getFlags = async () => {
-      const promise = new Promise((resolve) => {
-        setTimeout(() => {
-          resolve({
-            'profile-link': false,
-            profile: true,
-          });
-        }, 500);
-      });
-
-      const flags = await promise;
-
-      setFeatures({
-        flags,
-        loading: false,
-      });
-    };
-
-    getFlags();
-  }, []);
-
-  useEffect(() => {
-    const token = store.get();
-    if (token) {
-      setState(prevState => ({
-        ...prevState,
-        loggedIn: true,
-      }));
-    }
-
-    // TODO: This is temporary tracking to validate setup
-    track.pageview('/');
-    track.event({
-      category: 'Main',
-      action: 'Generic Action',
-    });
-  }, []);
 
   // performs action when user clicks the ROAM button
   const handleSearch = (destination) => {
@@ -89,6 +35,7 @@ function Home() {
       isLoading: true,
       destination,
     }));
+
     // request to backend that will request to API and send back the data
     axios.get(`${process.env.REACT_APP_ENDPOINT}/places/details/${destination}`)
       .then(({ data: { places } }) => {
@@ -113,66 +60,33 @@ function Home() {
       });
   };
 
-  const showModal = async (place) => {
-    setState(prevState => ({
-      ...prevState,
-      modal: {
-        ...prevState.modal,
-        show: true,
-        attraction: place,
-      },
-    }));
-  };
-
-  const closeModal = () => {
-    setState(prevState => ({
-      ...prevState,
-      modal: {
-        ...prevState.modal,
-        show: false,
-        attraction: {},
-      },
-    }));
-  };
-
-  const classTest = !state.modal.show ? styles.App : `${styles.App} ${styles.blur}`;
-
   return (
     <>
-      <div className={classTest}>
-        <Hero background="/images/hero.jpg">
-          {
-            !state.noResults && state.destination
-              ? <span className={styles.App__destination}>{state.destination}</span>
-              : <Search handleSearch={handleSearch} noResults={state.noResults} />
-          }
-        </Hero>
-        <Feature.Toggle flag="eslint">
-          <b>Eslint will fail without this until we actually implement a real toggle</b>
-        </Feature.Toggle>
+      <Hero background="/images/hero.jpg">
         {
-          !state.attractions.length > 0 && !state.isLoading
-            ? <PopularDestinations handleSearch={handleSearch} noResults={state.noResults} />
-            : (
-              <Attractions
-                attractions={state.attractions}
-                isLoading={state.isLoading}
-                showModal={showModal}
-              />
-            )
+          !state.noResults && state.destination
+            ? <span className={styles.Hero__basic}>{state.destination}</span>
+            : <Search handleSearch={handleSearch} noResults={state.noResults} />
         }
-      </div>
+      </Hero>
 
-      {state.modal.show && (
-        <Modal
-          attraction={state.modal.attraction}
-          onClose={closeModal}
-          showModal={showModal}
-          show={state.modal.show}
-        />
-      )}
+      {
+        !state.attractions.length > 0 && !state.isLoading
+          ? <PopularDestinations handleSearch={handleSearch} noResults={state.noResults} />
+          : (
+            <Attractions
+              attractions={state.attractions}
+              isLoading={state.isLoading}
+              showModal={showModal}
+            />
+          )
+      }
     </>
   );
 }
+
+Home.propTypes = {
+  showModal: PropTypes.func.isRequired,
+};
 
 export default Home;
